(SERVER-1773) Add support for bumping versions with qualifiers
This commit adds support to the plugin for bumping version numbers which
have a qualifier in them.  For example, with a version number of
5.4.0.0-3-SNAPSHOT, "3" would be the qualifier.  The default bump done
for a version with a qualifier would be a bump to the qualifier portion
only.  For example, 5.4.0.0-3-SNAPSHOT would be bumped to
5.4.0.0-4-SNAPSHOT.  The semantics for a qualifier follow the semantics
in the lein release plugin itself only with additional handling for
4-part version numbers that may have the additional qualifier.